# onyx-procurement/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ONYX Procurement API
  description: |
    API for the ONYX Procurement System — covers procurement (suppliers,
    RFQ, quotes, POs, subcontractors), VAT (PCN836), annual tax forms,
    bank reconciliation, payroll/wage slips, and WhatsApp webhooks.
  version: 1.0.0
  contact:
    name: Techno-Kol Ops
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://api.onyx.example.com
    description: Production placeholder
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    BankAccount:
      type: object
      additionalProperties: true
...
    BankTransaction:
      type: object
      additionalProperties: true
    ReconciliationMatch:
      type: object
      additionalProperties: true
...
  responses:
...
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/bank/accounts:
    get:
      tags: [Bank Reconciliation]
      summary: List bank accounts
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items: { $ref: "#/components/schemas/BankAccount" }
    post:
      tags: [Bank Reconciliation]
      summary: Create bank account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankAccount" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  account: { $ref: "#/components/schemas/BankAccount" }
  /api/bank/accounts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [Bank Reconciliation]
      summary: Update account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankAccount" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  account: { $ref: "#/components/schemas/BankAccount" }
  /api/bank/accounts/{id}/import:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Bank Reconciliation]
      summary: Import a bank statement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                format: { type: string }
                openingBalance: { type: number }
      responses:
        "201":
          description: Statement imported
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400": { $ref: "#/components/responses/BadRequest" }
        "422":
          description: Parse failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/bank/transactions:
    get:
      tags: [Bank Reconciliation]
      summary: List transactions
      parameters:
        - in: query
          name: account_id
          schema: { type: string }
        - in: query
          name: reconciled
          schema: { type: string, enum: ["true", "false"] }
        - in: query
          name: limit
          schema: { type: integer, default: 200 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items: { $ref: "#/components/schemas/BankTransaction" }
  /api/bank/accounts/{id}/auto-reconcile:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Bank Reconciliation]
      summary: Suggest matches for unreconciled transactions (read-only)
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/bank/matches:
    post:
      tags: [Bank Reconciliation]
      summary: Commit a reconciliation match
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReconciliationMatch" }
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                type: object
                properties:
                  match: { $ref: "#/components/schemas/ReconciliationMatch" }
  /api/bank/discrepancies:
    get:
      tags: [Bank Reconciliation]
      summary: List discrepancies
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...